requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pymongo==4.6.0
python-dotenv==1.0.0
python-multipart==0.0.12
pydantic[email]==2.9.2
requests==2.32.3

